requirements.txt
langchain
ipykernel
python-dotenv
langchain_community
pypdf
bs4
arxiv
pymupdf
wikipedia
langchain-text-splitters
langchain-openai
chromadb
sentence_transformers
langchain_huggingface
faiss-cpu
langchain_chroma
streamlit
langchain_groq
langchain_core
fastapi
langserve
sse_starlette
duckduckgo-search
mysql-connector-python
SQLAlchemy
validators
youtube_transcript_api
pytubefix
unstructured
langchain-yt-dlp
numexpr
huggingface_hub
gradio